chore: gatheringsDetailHandler 분리
